final clean & push b4 merge
Only thing that I need to figure out how to do is separate out all our subgraphs when we have multiple from our connection type. This is an edge case of the algo that is handled but how to parse.... Already tried a couple of things but am unsure, right now using something like a vector may be nice.