Odd behavior but working
For some reason Graph class randomly was getting mad due to not having <map> included but was working fine before. 

Also, in Node.hpp the way I iterate through our out edges (they are unique_ptrs) was wrong. I tried using a verbose naming to show that variables we are dealing with but sadly this did not compile. I think I have used that method before but for some reason it didnt work now. As of now I am doing something bad and using auto (we should never use auto due to current c++ coding paradigms, it used to be ok but not anymore)