Slight logic refactor
Switching over to something that will prevent us from the induced structure problem I mentioned. Had an incorrect understanding of how std::set is ordered, it doe NOT order itself temporally (when we insert an object). It uses either an user defined comparator or the default one. I was just getting lucky with where everything was placed in memory when testing, this would have caused some odd behavior but caught it before became an issue.

That being said, I use unordered_sets in our cycleDecomp and I use vectors (in order to get our last inserted node) in our subgraph matching. The end result is us getting "pairs" of values being the nodes we are worried about and the edges corresponding with how those nodes are connected.